+++ AmetrinTest/src/main/java/com/ametrinstudios/ametrin_test/data/provider/TestRecipeProvider.java
@@ -7,15 +7,15 @@ import net.minecraft.core.Registry;
 import net.minecraft.core.registries.MultiRegistryBootstrap;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
-import net.minecraft.resources.Identifier;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.crafting.Recipe;
 
+import java.util.Map;
 import java.util.Set;
 
 public final class TestRecipeProvider extends ExtendedRecipeProvider {
-    public TestRecipeProvider(BootstrapContext<Recipe<?>> provider, BootstrapContext<Advancement> output, Set<Identifier> recipeSet) {
-        super(AmetrinTestMod.MOD_ID, provider, output, recipeSet);
+    public TestRecipeProvider(BootstrapContext<Recipe<?>> recipeOutput, BootstrapContext<Advancement> advancementOutput) {
+        super(AmetrinTestMod.MOD_ID, recipeOutput, advancementOutput, Map.of());
     }
 
 
@@ -33,7 +33,7 @@ public final class TestRecipeProvider extends ExtendedRecipeProvider {
 
             @Override
             public void run(MultiRegistryBootstrap.BootstrapGetter registries) {
-                new TestRecipeProvider(registries.get(Registries.RECIPE), registries.get(Registries.ADVANCEMENT), Set.of()).buildRecipes();
+                new TestRecipeProvider(registries.get(Registries.RECIPE), registries.get(Registries.ADVANCEMENT)).buildRecipes();
             }
         };
     }