+++ changelog.md
@@ -1,5 +1,6 @@
 ## 1.2.0
 - add `ExtendedLootSubProvider` (provides common `HolderGetter`s)
+- `ExtendedRecipeProvider` takes a `Map<Block, BlockFamily>` and patches vanilla methods to work with custom families 
 - remove `CustomLootTableProvider` (use vanilla `LootTableProvider` directly)
 - remove `VanillaCompat.addFlowerPot`
 - disabled tree API (Vanilla tree generation should not be able to do the same)